Tests/LibWeb/Text/expected/element-get-bounding-client-rect-of-sticky.txt

Bounding Client Rect: top=50, left=10, width=500, height=57
